# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AnxietyWatch API
  version: 0.1.0
  description: >-
    API de bienestar, monitoreo y apoyo. No diagnostica trastornos ni confirma
    clínicamente eventos de ansiedad.
servers:
  - url: http://localhost:3000
    description: Desarrollo local
paths:
  /health:
    get:
      operationId: getHealth
      summary: Comprueba el estado del API
      responses:
        '200':
          description: El servicio está disponible
          headers:
            x-correlation-id:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'
  /api/v1/telemetry/batch:
    post:
      operationId: postTelemetryBatch
      summary: Recibe un lote de telemetria del movil
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelemetryBatch'
      responses:
        '202': { description: Lote aceptado }
        '200': { description: Lote duplicado }
        '400': { description: Payload invalido }
  /api/v1/sos/trigger:
    post:
      operationId: postSosTrigger
      summary: Activa una alerta SOS idempotente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SosTrigger'
      responses:
        '202': { description: Evento aceptado }
        '200': { description: Evento duplicado }
        '400': { description: Payload invalido }
  /api/v1/sos/cancel:
    post:
      operationId: postSosCancel
      summary: Cancela una alerta SOS activa, idempotente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SosCancel'
      responses:
        '202': { description: Cancelacion aceptada }
        '200': { description: Cancelacion duplicada }
        '400': { description: Payload invalido }
components:
  schemas:
    TelemetryBatch:
      type: object
      additionalProperties: false
      required:
        [
          batchId,
          deviceId,
          userId,
          sessionId,
          startedAt,
          endedAt,
          sequence,
          samples,
        ]
      properties:
        batchId: { type: string, format: uuid }
        deviceId: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        sessionId: { type: string, format: uuid }
        startedAt: { type: string, format: date-time }
        endedAt: { type: string, format: date-time }
        sequence: { type: integer, minimum: 0 }
        samples:
          type: array
          minItems: 1
          maxItems: 600
          items:
            type: object
    SosTrigger:
      type: object
      additionalProperties: false
      required: [eventId, userId, deviceId, triggeredAt, source]
      properties:
        eventId: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        deviceId: { type: string, format: uuid }
        triggeredAt: { type: string, format: date-time }
        source: { type: string, enum: [WATCH, MOBILE] }
        reason: { type: [string, 'null'], maxLength: 500 }
    SosCancel:
      type: object
      additionalProperties: false
      required: [eventId, userId, deviceId, cancelledAt]
      properties:
        eventId: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        deviceId: { type: string, format: uuid }
        cancelledAt: { type: string, format: date-time }
        reason: { type: [string, 'null'], maxLength: 500 }
    HealthStatus:
      type: object
      additionalProperties: false
      required: [status, service, version]
      properties:
        status:
          type: string
          const: ok
        service:
          type: string
        version:
          type: string
    SensorAvailability:
      type: string
      enum: [available, unsupported, unavailable]
    DeviceCapabilities:
      type: object
      additionalProperties: false
      required:
        [deviceModel, wearOsVersion, healthPlatformVersion, capabilities]
      properties:
        deviceModel:
          type: string
        wearOsVersion:
          type: string
        healthPlatformVersion:
          type: [string, 'null']
        capabilities:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/SensorAvailability'